Fixed deleting a calendar before its first sync. Disabled retry for sync queue messages.

+++ FastiCalSync.Functions/SyncJobProcessor.cs
@@ -63,8 +63,11 @@ namespace FastiCalSync.Functions
 
             if (calendar.SyncState == SyncState.Deleting)
             {
-                log.Info("Deleting Google calendar...");
-                await gCalService.DeleteCalendar(calendar.GoogleCalendarID);
+                if (calendar.GoogleCalendarID != null)
+                {
+                    log.Info("Deleting Google calendar...");
+                    await gCalService.DeleteCalendar(calendar.GoogleCalendarID);
+                }
 
                 log.Info("Deleting calendar record...");
                 await calendarRepository.Delete(calendar);
